change of schema for conflict list and structure of the data in the mongoDB
- removed the properties from the mongoDB collections, since the
references to the property definitions might be altered by a tool; it is
just an ID

- made the decision to handle additional properties explicitly and
therefore independent of the storage of a particular node or relation

- can not create yet a conflict, because the current export of the
exchange format from Archi does not export the version attribute of the
root element. As a consequence the version number is empty and therefore
something goes wrong in the processJsonString function. To be fixed!